+++ lib/generated_plugin_registrant.dart
@@ -5,6 +5,7 @@
 // ignore: unused_import
 import 'dart:ui';
 
+import 'package:connectivity_for_web/connectivity_for_web.dart';
 import 'package:shared_preferences_web/shared_preferences_web.dart';
 import 'package:url_launcher_web/url_launcher_web.dart';
 
@@ -12,6 +13,7 @@ import 'package:flutter_web_plugins/flutter_web_plugins.dart';
 
 // ignore: public_member_api_docs
 void registerPlugins(PluginRegistry registry) {
+  ConnectivityPlugin.registerWith(registry.registrarFor(ConnectivityPlugin));
   SharedPreferencesPlugin.registerWith(registry.registrarFor(SharedPreferencesPlugin));
   UrlLauncherPlugin.registerWith(registry.registrarFor(UrlLauncherPlugin));
   registry.registerMessageHandler();